requirements.txt
pytest
networkx
pandas==0.23.4
numpy==1.17.4
scikit-learn==0.20.3
pyyaml==5.1
numba==0.43.1
python-levenshtein
seaborn
matplotlib
line_profiler
jellyfish
Fuzzy
pyjarowinkler
editdistance
street-address
nmslib==2.0.5
prompt_toolkit
pyarrow
Dickens
llvmlite==0.28.0
git+https://github.com/mmcneill/memory_profiler.git@fdf4488ffe42c588bfa632537e9a959e4b36bf83#egg=memory_profiler
git+https://github.com/zjelveh/name-probability.git@64e64d2948a860802fa3401cc2b3c897a5abdb2a#egg=NameProbability
